List files and directories in GitLab repository

Lists file and directory names found under the specified path in a GitLab repository via the GET /projects/:id/repository/tree endpoint of GitLab's v4 RESTful API.

Usage

dir_ls(
  path = "",
  id_project = pal::pkg_config_val("id_project"),
  recurse = FALSE,
  ref = project_default_branch(id_project = id_project, base_url = base_url, token =
    token, max_tries = max_tries),
  base_url = pal::pkg_config_val("base_url"),
  token = pal::pkg_config_val("token"),
  max_tries = 3L
)

Arguments

path

character(1)
Directory path, relative to the repository root.

id_project

integer(1)
GitLab project identifier.

recurse

logical(1)
Whether or not to recurse into subdirectories of path.

ref

character(1)
Git revision expression matching the desired Git tree object, e.g. a ref name (branch, tag, etc.), a commit identifier, or another symbolic reference like "HEAD~10". Note that the GitLab API doesn't support every type of revision expression. Defaults to "HEAD" if NULL.

base_url

character(1)
Base URL to the GitLab v4 RESTful API root.

token

character(1)
GitLab access token used for authentication. If NULL, the API request is made unauthenticated which results in HTTP 404 Not Found and other errors for non-public resources.

max_tries

integer(1)
Maximum number of attempts to retry in case of an HTTP error.

Value

A tibble.

See also

Examples

gitlab::dir_ls(id_project = 64767928,
               path = "man",
               recurse = TRUE)
#> # A tibble: 21 × 5
#>    id                                       name                  type  path                      mode  
#>    <chr>                                    <chr>                 <chr> <chr>                     <chr> 
#>  1 0eac40f6eba506fdd4e71c3ecd05cbc96f886598 roxygen               tree  man/roxygen               040000
#>  2 9ec6cc14128019433014cc49c9cb70d729c50ebb api_req.Rd            blob  man/api_req.Rd            100644
#>  3 b8082bc2a9e4591aad90304f9d22bfceaafe5e6c build_url.Rd          blob  man/build_url.Rd          100644
#>  4 5daa8ce07529bcd2904b4ff6528b266b6d32c353 commit_files.Rd       blob  man/commit_files.Rd       100644
#>  5 1f2838ca53f25e157e7601bc385059d1f99a1dae dir_delete.Rd         blob  man/dir_delete.Rd         100644
#>  6 04ce8bc217ca094614708cc5cd0f7880a2f0322f dir_ls.Rd             blob  man/dir_ls.Rd             100644
#>  7 1d632247987917b16b93e981eb2092d6f4dbc4aa file_commit_action.Rd blob  man/file_commit_action.Rd 100644
#>  8 7f6a1cf38c9c7642126a75c2e1c853fc7e2a4512 file_content.Rd       blob  man/file_content.Rd       100644
#>  9 1cb23be95425a4da228ac460013a7e0b469f116e file_create.Rd        blob  man/file_create.Rd        100644
#> 10 e885012ddc598c9ff923837e6e338211aa2c7238 file_delete.Rd        blob  man/file_delete.Rd        100644
#> # ℹ 11 more rows