GitLab project metadata

Retrieves a GitLab project's metadata via the GET /projects/{id_project} endpoint of GitLab's v4 RESTful API.

Usage

project(
  id_project = pal::pkg_config_val("id_project"),
  base_url = pal::pkg_config_val("base_url"),
  token = pal::pkg_config_val("token"),
  max_tries = 3L
)

Arguments

id_project

integer(1)
GitLab project identifier.

base_url

character(1)
Base URL to the GitLab v4 RESTful API root.

token

character(1)
GitLab access token used for authentication. If NULL, the API request is made unauthenticated which results in HTTP 404 Not Found and other errors for non-public resources.

max_tries

integer(1)
Maximum number of attempts to retry in case of an HTTP error.

Value

A list.

See also

Examples

gitlab::project(id_project = 64767928)
#> $id
#> [1] 64767928
#> 
#> $description
#> [1] "**gitlab** is an [R](https://www.r-project.org/) [📦](https://r-pkgs.org/) that wraps [GitLab's v4 RESTful API](https://docs.gitlab.com/ee/api/rest/)."
#> 
#> $name
#> [1] "gitlab"
#> 
#> $name_with_namespace
#> [1] "rpkg.dev / gitlab"
#> 
#> $path
#> [1] "gitlab"
#> 
#> $path_with_namespace
#> [1] "rpkg.dev/gitlab"
#> 
#> $created_at
#> [1] "2024-11-22T21:52:01.779Z"
#> 
#> $default_branch
#> [1] "main"
#> 
#> $tag_list
#> list()
#> 
#> $topics
#> list()
#> 
#> $ssh_url_to_repo
#> [1] "git@gitlab.com:rpkg.dev/gitlab.git"
#> 
#> $http_url_to_repo
#> [1] "https://gitlab.com/rpkg.dev/gitlab.git"
#> 
#> $web_url
#> [1] "https://gitlab.com/rpkg.dev/gitlab"
#> 
#> $readme_url
#> [1] "https://gitlab.com/rpkg.dev/gitlab/-/blob/main/README.Rmd"
#> 
#> $forks_count
#> [1] 0
#> 
#> $avatar_url
#> NULL
#> 
#> $star_count
#> [1] 0
#> 
#> $last_activity_at
#> [1] "2024-11-24T23:26:01.620Z"
#> 
#> $namespace
#> $namespace$id
#> [1] 11130607
#> 
#> $namespace$name
#> [1] "rpkg.dev"
#> 
#> $namespace$path
#> [1] "rpkg.dev"
#> 
#> $namespace$kind
#> [1] "group"
#> 
#> $namespace$full_path
#> [1] "rpkg.dev"
#> 
#> $namespace$parent_id
#> NULL
#> 
#> $namespace$avatar_url
#> NULL
#> 
#> $namespace$web_url
#> [1] "https://gitlab.com/groups/rpkg.dev"
#> 
#> 
#> $container_registry_image_prefix
#> [1] "registry.gitlab.com/rpkg.dev/gitlab"
#> 
#> $`_links`
#> $`_links`$self
#> [1] "https://gitlab.com/api/v4/projects/64767928"
#> 
#> $`_links`$issues
#> [1] "https://gitlab.com/api/v4/projects/64767928/issues"
#> 
#> $`_links`$merge_requests
#> [1] "https://gitlab.com/api/v4/projects/64767928/merge_requests"
#> 
#> $`_links`$repo_branches
#> [1] "https://gitlab.com/api/v4/projects/64767928/repository/branches"
#> 
#> $`_links`$labels
#> [1] "https://gitlab.com/api/v4/projects/64767928/labels"
#> 
#> $`_links`$events
#> [1] "https://gitlab.com/api/v4/projects/64767928/events"
#> 
#> $`_links`$members
#> [1] "https://gitlab.com/api/v4/projects/64767928/members"
#> 
#> $`_links`$cluster_agents
#> [1] "https://gitlab.com/api/v4/projects/64767928/cluster_agents"
#> 
#> 
#> $packages_enabled
#> [1] TRUE
#> 
#> $empty_repo
#> [1] FALSE
#> 
#> $archived
#> [1] FALSE
#> 
#> $visibility
#> [1] "public"
#> 
#> $resolve_outdated_diff_discussions
#> [1] FALSE
#> 
#> $container_expiration_policy
#> $container_expiration_policy$cadence
#> [1] "1d"
#> 
#> $container_expiration_policy$enabled
#> [1] FALSE
#> 
#> $container_expiration_policy$keep_n
#> [1] 10
#> 
#> $container_expiration_policy$older_than
#> [1] "90d"
#> 
#> $container_expiration_policy$name_regex
#> [1] ".*"
#> 
#> $container_expiration_policy$name_regex_keep
#> NULL
#> 
#> $container_expiration_policy$next_run_at
#> [1] "2024-11-23T21:52:01.813Z"
#> 
#> 
#> $repository_object_format
#> [1] "sha1"
#> 
#> $issues_enabled
#> [1] TRUE
#> 
#> $merge_requests_enabled
#> [1] TRUE
#> 
#> $wiki_enabled
#> [1] TRUE
#> 
#> $jobs_enabled
#> [1] TRUE
#> 
#> $snippets_enabled
#> [1] TRUE
#> 
#> $container_registry_enabled
#> [1] TRUE
#> 
#> $service_desk_enabled
#> [1] TRUE
#> 
#> $service_desk_address
#> [1] "contact-project+rpkg-dev-gitlab-64767928-issue-@incoming.gitlab.com"
#> 
#> $can_create_merge_request_in
#> [1] TRUE
#> 
#> $issues_access_level
#> [1] "enabled"
#> 
#> $repository_access_level
#> [1] "enabled"
#> 
#> $merge_requests_access_level
#> [1] "enabled"
#> 
#> $forking_access_level
#> [1] "enabled"
#> 
#> $wiki_access_level
#> [1] "enabled"
#> 
#> $builds_access_level
#> [1] "enabled"
#> 
#> $snippets_access_level
#> [1] "enabled"
#> 
#> $pages_access_level
#> [1] "private"
#> 
#> $analytics_access_level
#> [1] "enabled"
#> 
#> $container_registry_access_level
#> [1] "enabled"
#> 
#> $security_and_compliance_access_level
#> [1] "private"
#> 
#> $releases_access_level
#> [1] "enabled"
#> 
#> $environments_access_level
#> [1] "enabled"
#> 
#> $feature_flags_access_level
#> [1] "enabled"
#> 
#> $infrastructure_access_level
#> [1] "enabled"
#> 
#> $monitor_access_level
#> [1] "enabled"
#> 
#> $model_experiments_access_level
#> [1] "enabled"
#> 
#> $model_registry_access_level
#> [1] "enabled"
#> 
#> $emails_disabled
#> [1] FALSE
#> 
#> $emails_enabled
#> [1] TRUE
#> 
#> $shared_runners_enabled
#> [1] TRUE
#> 
#> $lfs_enabled
#> [1] TRUE
#> 
#> $creator_id
#> [1] 1089770
#> 
#> $import_url
#> NULL
#> 
#> $import_type
#> NULL
#> 
#> $import_status
#> [1] "none"
#> 
#> $import_error
#> NULL
#> 
#> $open_issues_count
#> [1] 0
#> 
#> $description_html
#> [1] "<p data-sourcepos=\"1:1-1:152\" dir=\"auto\"><strong data-sourcepos=\"1:1-1:10\">gitlab</strong> is an <a data-sourcepos=\"1:18-1:48\" href=\"https://www.r-project.org/\" rel=\"nofollow noreferrer noopener\" target=\"_blank\">R</a> <a data-sourcepos=\"1:50-1:76\" href=\"https://r-pkgs.org/\" rel=\"nofollow noreferrer noopener\" target=\"_blank\"><gl-emoji title=\"package\" data-name=\"package\" data-unicode-version=\"6.0\">📦</gl-emoji></a> that wraps <a data-sourcepos=\"1:89-1:151\" href=\"https://docs.gitlab.com/ee/api/rest/\" rel=\"nofollow noreferrer noopener\" target=\"_blank\">GitLab's v4 RESTful API</a>.</p>"
#> 
#> $updated_at
#> [1] "2024-11-24T23:41:16.163Z"
#> 
#> $ci_default_git_depth
#> [1] 20
#> 
#> $ci_forward_deployment_enabled
#> [1] TRUE
#> 
#> $ci_forward_deployment_rollback_allowed
#> [1] TRUE
#> 
#> $ci_job_token_scope_enabled
#> [1] FALSE
#> 
#> $ci_separated_caches
#> [1] TRUE
#> 
#> $ci_allow_fork_pipelines_to_run_in_parent_project
#> [1] TRUE
#> 
#> $ci_id_token_sub_claim_components
#> $ci_id_token_sub_claim_components[[1]]
#> [1] "project_path"
#> 
#> $ci_id_token_sub_claim_components[[2]]
#> [1] "ref_type"
#> 
#> $ci_id_token_sub_claim_components[[3]]
#> [1] "ref"
#> 
#> 
#> $build_git_strategy
#> [1] "fetch"
#> 
#> $keep_latest_artifact
#> [1] TRUE
#> 
#> $restrict_user_defined_variables
#> [1] FALSE
#> 
#> $ci_pipeline_variables_minimum_override_role
#> [1] "maintainer"
#> 
#> $runners_token
#> NULL
#> 
#> $runner_token_expiration_interval
#> NULL
#> 
#> $group_runners_enabled
#> [1] TRUE
#> 
#> $auto_cancel_pending_pipelines
#> [1] "enabled"
#> 
#> $build_timeout
#> [1] 3600
#> 
#> $auto_devops_enabled
#> [1] FALSE
#> 
#> $auto_devops_deploy_strategy
#> [1] "continuous"
#> 
#> $ci_push_repository_for_job_token_allowed
#> [1] FALSE
#> 
#> $ci_config_path
#> [1] ""
#> 
#> $public_jobs
#> [1] TRUE
#> 
#> $shared_with_groups
#> list()
#> 
#> $only_allow_merge_if_pipeline_succeeds
#> [1] FALSE
#> 
#> $allow_merge_on_skipped_pipeline
#> NULL
#> 
#> $request_access_enabled
#> [1] TRUE
#> 
#> $only_allow_merge_if_all_discussions_are_resolved
#> [1] FALSE
#> 
#> $remove_source_branch_after_merge
#> [1] TRUE
#> 
#> $printing_merge_request_link_enabled
#> [1] TRUE
#> 
#> $merge_method
#> [1] "merge"
#> 
#> $squash_option
#> [1] "default_off"
#> 
#> $enforce_auth_checks_on_uploads
#> [1] TRUE
#> 
#> $suggestion_commit_message
#> NULL
#> 
#> $merge_commit_template
#> NULL
#> 
#> $squash_commit_template
#> NULL
#> 
#> $issue_branch_template
#> NULL
#> 
#> $warn_about_potentially_unwanted_characters
#> [1] TRUE
#> 
#> $autoclose_referenced_issues
#> [1] TRUE
#> 
#> $external_authorization_classification_label
#> [1] ""
#> 
#> $requirements_enabled
#> [1] FALSE
#> 
#> $requirements_access_level
#> [1] "enabled"
#> 
#> $security_and_compliance_enabled
#> [1] TRUE
#> 
#> $compliance_frameworks
#> list()
#> 
#> $permissions
#> $permissions$project_access
#> NULL
#> 
#> $permissions$group_access
#> $permissions$group_access$access_level
#> [1] 50
#> 
#> $permissions$group_access$notification_level
#> [1] 2
#> 
#> 
#>