List files and directories in GitLab repository

Lists file and directory names found under the specified path in a GitLab repository via the GET /projects/:id/repository/tree endpoint of GitLab's v4 RESTful API.

Usage

dir_ls(
  id_proj,
  path = "",
  recursive = FALSE,
  ref = project_default_branch(id_proj = id_proj, base_url = base_url, token = token,
    max_tries = max_tries),
  base_url = pal::pkg_config_val("base_url"),
  token = pal::pkg_config_val("token"),
  max_tries = 3L
)

Arguments

id_proj

integer(1)
GitLab project identifier.

path

character(1)
Directory path, relative to the repository root.

recursive

logical(1)
Whether or not to recurse into subdirectories of path.

ref

character(1)
Git revision expression matching the desired Git tree object, e.g. a ref name (branch, tag, etc.), a commit identifier, or another symbolic reference like "HEAD~10". Note that the GitLab API doesn't support every type of revision expression. Defaults to "HEAD" if NULL.

base_url

character(1)
Base URL to the GitLab v4 RESTful API root.

token

character(1)
GitLab access token used for authentication. If NULL, the API request is made unauthenticated which results in HTTP 404 Not Found and other errors for non-public resources.

max_tries

integer(1)
Maximum number of attempts to retry in case of an HTTP error.

Value

A tibble.

Examples

gitlab::dir_ls(id_proj = 64767928,
               path = "man",
               recursive = TRUE)
#> # A tibble: 17 × 5
#>    id                                       name                      type  path                          mode  
#>    <chr>                                    <chr>                     <chr> <chr>                         <chr> 
#>  1 6dc958d6414dd9a2ee7f59876823711e09cbdf0d roxygen                   tree  man/roxygen                   040000
#>  2 9ec6cc14128019433014cc49c9cb70d729c50ebb api_req.Rd                blob  man/api_req.Rd                100644
#>  3 33f5ecf9b6bf8d7caa2ef1fabdfd3a0130165fc1 build_url.Rd              blob  man/build_url.Rd              100644
#>  4 d75193fa07313559f7dc146c7a27a9a4e6e2f04e dir_ls.Rd                 blob  man/dir_ls.Rd                 100644
#>  5 1bc528df4eadf1eda2134a020479054ca1f658b5 file_content.Rd           blob  man/file_content.Rd           100644
#>  6 306f0961cea3ec92b8100e7fe93f7b6ff06314a8 file_create.Rd            blob  man/file_create.Rd            100644
#>  7 bab487d2bd22a8a4ebfa913ec4092baedcf385fa file_delete.Rd            blob  man/file_delete.Rd            100644
#>  8 b9565b8280dedd10c3fc2585d67b07dfb4ab687e file_exists.Rd            blob  man/file_exists.Rd            100644
#>  9 b4803cfd4fd542c8eaa91bc853d94117a9e4cba2 file_full.Rd              blob  man/file_full.Rd              100644
#> 10 996b310b0a48384c42577ed9dc1d95bdd75bd44c file_meta.Rd              blob  man/file_meta.Rd              100644
#> 11 67986dc18b4d625e0a61ee62ba15846b840119b6 file_req.Rd               blob  man/file_req.Rd               100644
#> 12 8b2356a76628a46fb59b19c725cc22916362ddae file_update.Rd            blob  man/file_update.Rd            100644
#> 13 7dd90d49713fba0f2c9b6aac93b93112e9e51e81 file_write.Rd             blob  man/file_write.Rd             100644
#> 14 bea606fd667ef4aa2369962ada62c806cfd10953 gitlab-package.Rd         blob  man/gitlab-package.Rd         100644
#> 15 23bb4905235de70fe036a97c19343a6afb1bd46d project.Rd                blob  man/project.Rd                100644
#> 16 28f41c020ce4be39d425fd1b9ea91cbf1cb43d25 project_default_branch.Rd blob  man/project_default_branch.Rd 100644
#> 17 bbda19aaf6dbe500ef344a79b327dcf66953587f meta.R                    blob  man/roxygen/meta.R            100644