GitLab project metadata

Retrieves a GitLab project's metadata via the GET /projects/{id_project} endpoint of GitLab's v4 RESTful API.

Usage

project(
  id_project = pal::pkg_config_val("id_project"),
  base_url = pal::pkg_config_val("base_url"),
  token = pal::pkg_config_val("token"),
  max_tries = 3L
)

Arguments

id_project

integer(1)
GitLab project identifier.

base_url

character(1)
Base URL to the GitLab v4 RESTful API root.

token

character(1)
GitLab access token used for authentication. If NULL, the API request is made unauthenticated which results in HTTP 404 Not Found and other errors for non-public resources.

max_tries

integer(1)
Maximum number of attempts to retry in case of an HTTP error.

Value

A list.

See also

Examples

gitlab::project(id_project = 64767928) |> names()
#>   [1] "id"                                               "description"                                      "name"                                            
#>   [4] "name_with_namespace"                              "path"                                             "path_with_namespace"                             
#>   [7] "created_at"                                       "default_branch"                                   "tag_list"                                        
#>  [10] "topics"                                           "ssh_url_to_repo"                                  "http_url_to_repo"                                
#>  [13] "web_url"                                          "readme_url"                                       "forks_count"                                     
#>  [16] "avatar_url"                                       "star_count"                                       "last_activity_at"                                
#>  [19] "namespace"                                        "container_registry_image_prefix"                  "_links"                                          
#>  [22] "packages_enabled"                                 "empty_repo"                                       "archived"                                        
#>  [25] "visibility"                                       "resolve_outdated_diff_discussions"                "container_expiration_policy"                     
#>  [28] "repository_object_format"                         "issues_enabled"                                   "merge_requests_enabled"                          
#>  [31] "wiki_enabled"                                     "jobs_enabled"                                     "snippets_enabled"                                
#>  [34] "container_registry_enabled"                       "service_desk_enabled"                             "service_desk_address"                            
#>  [37] "can_create_merge_request_in"                      "issues_access_level"                              "repository_access_level"                         
#>  [40] "merge_requests_access_level"                      "forking_access_level"                             "wiki_access_level"                               
#>  [43] "builds_access_level"                              "snippets_access_level"                            "pages_access_level"                              
#>  [46] "analytics_access_level"                           "container_registry_access_level"                  "security_and_compliance_access_level"            
#>  [49] "releases_access_level"                            "environments_access_level"                        "feature_flags_access_level"                      
#>  [52] "infrastructure_access_level"                      "monitor_access_level"                             "model_experiments_access_level"                  
#>  [55] "model_registry_access_level"                      "emails_disabled"                                  "emails_enabled"                                  
#>  [58] "shared_runners_enabled"                           "lfs_enabled"                                      "creator_id"                                      
#>  [61] "import_url"                                       "import_type"                                      "import_status"                                   
#>  [64] "import_error"                                     "open_issues_count"                                "description_html"                                
#>  [67] "updated_at"                                       "ci_default_git_depth"                             "ci_forward_deployment_enabled"                   
#>  [70] "ci_forward_deployment_rollback_allowed"           "ci_job_token_scope_enabled"                       "ci_separated_caches"                             
#>  [73] "ci_allow_fork_pipelines_to_run_in_parent_project" "ci_id_token_sub_claim_components"                 "build_git_strategy"                              
#>  [76] "keep_latest_artifact"                             "restrict_user_defined_variables"                  "ci_pipeline_variables_minimum_override_role"     
#>  [79] "runners_token"                                    "runner_token_expiration_interval"                 "group_runners_enabled"                           
#>  [82] "auto_cancel_pending_pipelines"                    "build_timeout"                                    "auto_devops_enabled"                             
#>  [85] "auto_devops_deploy_strategy"                      "ci_push_repository_for_job_token_allowed"         "ci_config_path"                                  
#>  [88] "public_jobs"                                      "shared_with_groups"                               "only_allow_merge_if_pipeline_succeeds"           
#>  [91] "allow_merge_on_skipped_pipeline"                  "request_access_enabled"                           "only_allow_merge_if_all_discussions_are_resolved"
#>  [94] "remove_source_branch_after_merge"                 "printing_merge_request_link_enabled"              "merge_method"                                    
#>  [97] "squash_option"                                    "enforce_auth_checks_on_uploads"                   "suggestion_commit_message"                       
#> [100] "merge_commit_template"                            "squash_commit_template"                           "issue_branch_template"                           
#> [103] "warn_about_potentially_unwanted_characters"       "autoclose_referenced_issues"                      "external_authorization_classification_label"     
#> [106] "requirements_enabled"                             "requirements_access_level"                        "security_and_compliance_enabled"                 
#> [109] "compliance_frameworks"                            "permissions"